+++ client/src/api/Years.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 import { Photo, PhotoCollection } from "../models/Photo";
 
 export async function fecthAll(): Promise<PhotoCollection[]> {
-  const res = await axios.get('/api/timeline');
+  const res = await axios.get('/api/years');
   const data = res.data;
   const byYears = data.years;
   return byYears.map(
