+++ routes/api/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 router.use('/albums', require('./album'));
-router.use('/timeline', require('./timeline'));
+router.use('/years', require('./years'));
 router.use('/media', require('./media'));
 
 module.exports = router;
